+++ platformer/Assets/PixelCrew/UI/Widgets/LifeBarWidget.cs
@@ -18,7 +18,7 @@ namespace PixelCrew.UI.Widgets
             if (_hp == null)
                 _hp = GetComponentInParent<HealthComponent>();
 
-            _maxHp = _hp.Health;
+            _maxHp = (int) _hp.Health;
 
            _trash.Retain(_hp._onDie.Subscribe(OnDie));
            _trash.Retain(_hp._onChange.Subscribe(OnHpChanged));
